Move PDF functionality into a submodule
