+++ src/error.rs
@@ -1,4 +1,5 @@
 use std::{io, result};
+use quick_error::quick_error;
 
 quick_error! {
     /// The error type for wkhtmltopdf generation
